fix(H-8): compute conformational Shannon entropy correctly
run_shannon_thermo_stack() built a distribution of -log(w_i) values, then
fed those into compute_shannon_entropy() which histogrammed them and
returned the entropy of the binned log-weight distribution. That quantity
has no thermodynamic interpretation — it is sensitive to bin count and
does not converge to the Boltzmann ensemble entropy as N→∞.

The correct conformational entropy of the Boltzmann distribution is
    H = -Σ_i w_i · ln(w_i)             (nats)
    S_conf = k_B · H                    (kcal/mol·K)

Replace the histogram path with a direct sum over the normalized
Boltzmann weights returned by StatMechEngine::boltzmann_weights().

+++ LIB/ParallelDock.cpp
@@ -248,7 +248,7 @@ statmech::StatMechEngine ParallelDockManager::get_global_engine() const {
         if (r.energies.empty()) continue;
         global.merge_samples(
             std::span<const double>(r.energies),
-            std::span<const int>(r.multiplicities)
+            std::span<const double>(r.multiplicities)
         );
     }
